refactor(ocale): rename place-publishing domain to universes
Renames domains/place-publishing/places/ to domains/universes/places/.
The /universes/v1/ path prefix maps to domains/universes/ under
ADR-022's mechanical URL-to-folder rule (drop the version, take the
prefix). The earlier place-publishing/ name was invented from
/cloud/guides/usage-place-publishing.md to avoid a perceived collision
with domains/cloud-v2/universes/, but those two paths play different
roles at different depths: one is a top-level Domain folder for a
URL prefix, the other is a per-shape sub-tree inside the cloud-v2
Domain. ADR-022 explicitly establishes that two-level structure, so
the same name appearing at different levels is the intended shape,
not a clash.

universes/ also leaves natural room for additional /universes/v1/
operations Roblox may add while the surface is BETA, and self-cleans
when those operations migrate to /cloud/v2/.

Imports updated in src/resources/places/client.ts and index.ts and in
tests/helpers/places.ts. Two descriptive comments that referenced the
old folder name (cloud-v2/places/operations.ts and the renamed
universes/places/wire.ts) now describe the operation directly. Wire
requests, parsed shapes, and operation keys (places.publishVersion,
places.update) are unchanged; all 1328 ocale tests pass and lint is
clean.
