+++ packages/open-cloud/src/domains/universes/places/signatures.ts
